+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * Function.series forces asynchronicity by default (by using setImmediate),
   but if wanted this can be disabled by passing a boolean false as first arg
 * Added shim for ES6's Array#fill(value, start, end)
+* Added Function.inherits, Function#extend and Function#setMethod
 
 ## 0.1.2 (2014-08-28)